pyproject.toml
[project]
name = "repo-explainer"
version = "0.1.0"
description = "A FastAPI backend that explains GitHub repositories using Claude AI"
authors = [
    { name = "Bao Nguyen", email = "thienbao05@arizona.edu" }
]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

# Core dependencies needed to run the application
dependencies = [
    "fastapi[standard]>=0.115.0",
    "anthropic>=0.40.0",
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
# Development tools
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "black>=24.0.0",
    "ruff>=0.8.0",  # Fast linter
    "mypy>=1.13.0",
    "ipython>=8.30.0",
]

# Testing tools
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.0",  # For testing async HTTP
]

[build-system]
requires = ["setuptools>=67.0.0", "wheel"]
build-backend = "setuptools.build_meta"

# Black configuration for code formatting
[tool.black]
line-length = 88
target-version = ['py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | \.pytest_cache
  | __pycache__
  | dist
)/
'''

# Ruff configuration (fast Python linter)
[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

# Pytest configuration
[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-v --cov=backend --cov-report=html --cov-report=term"
testpaths = ["backend/tests"]
pythonpath = ["."]
asyncio_mode = "auto"

# MyPy configuration for type checking
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Set to true when you're ready for strict typing
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.setuptools.packages.find]
include = ["backend*"]